refactor: Перенёс dto
