+++ ewm-main-service/src/main/java/ru/practicum/event/dto/EventRequestStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package ru.practicum.dto;
+package ru.practicum.event.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
