+++ ewm-main-service/src/main/java/ru/practicum/event/dto/EventSort.java
@@ -1,4 +1,4 @@
-package ru.practicum.dto.event;
+package ru.practicum.event.dto;
 
 /**
  * Список сортировок.
