+++ ewm-main-service/src/main/java/ru/practicum/event/dto/EventState.java
@@ -1,4 +1,4 @@
-package ru.practicum.dto.event;
+package ru.practicum.event.dto;
 
 /**
  * Список состояний жизненного цикла события.
