+++ ewm-main-service/src/main/java/ru/practicum/event/dto/StateReview.java
@@ -1,4 +1,4 @@
-package ru.practicum.dto.event;
+package ru.practicum.event.dto;
 
 /**
  * Список состояния ревью события.
