+++ ewm-main-service/src/main/java/ru/practicum/event/service/PrivateEventService.java
@@ -1,12 +1,12 @@
 package ru.practicum.event.service;
 
-import ru.practicum.dto.EventRequestStatusUpdateRequest;
-import ru.practicum.dto.EventRequestStatusUpdateResult;
+import ru.practicum.event.dto.EventRequestStatusUpdateRequest;
+import ru.practicum.event.dto.EventRequestStatusUpdateResult;
 import ru.practicum.participation.dto.ParticipationRequestDto;
-import ru.practicum.dto.event.EventDtoRequest;
-import ru.practicum.dto.event.EventFullDtoResponse;
-import ru.practicum.dto.event.EventShortDto;
-import ru.practicum.dto.event.UpdateEventUserRequest;
+import ru.practicum.event.dto.EventDtoRequest;
+import ru.practicum.event.dto.EventFullDtoResponse;
+import ru.practicum.event.dto.EventShortDto;
+import ru.practicum.event.dto.UpdateEventUserRequest;
 
 import java.util.List;
