+++ ewm-main-service/src/main/java/ru/practicum/event/service/impl/AdminEventServiceImpl.java
@@ -7,10 +7,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import ru.practicum.category.model.Category;
-import ru.practicum.dto.event.EventFullDtoResponse;
-import ru.practicum.dto.event.EventPublisedState;
-import ru.practicum.dto.event.EventState;
-import ru.practicum.dto.event.UpdateEventAdminRequest;
+import ru.practicum.event.dto.EventFullDtoResponse;
+import ru.practicum.event.dto.EventPublisedState;
+import ru.practicum.event.dto.EventState;
+import ru.practicum.event.dto.UpdateEventAdminRequest;
 import ru.practicum.event.mapper.EventMapper;
 import ru.practicum.event.model.Event;
 import ru.practicum.event.model.Location;
