+++ ewm-main-service/src/main/java/ru/practicum/event/service/impl/PrivateEventServiceImpl.java
@@ -7,10 +7,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.category.model.Category;
 import ru.practicum.category.service.CategoryService;
-import ru.practicum.dto.EventRequestStatusUpdateRequest;
-import ru.practicum.dto.EventRequestStatusUpdateResult;
+import ru.practicum.event.dto.*;
 import ru.practicum.participation.dto.ParticipationRequestDto;
-import ru.practicum.dto.event.*;
 import ru.practicum.event.mapper.EventMapper;
 import ru.practicum.event.repository.EventRepository;
 import ru.practicum.event.service.EventService;
