+++ ewm-main-service/src/main/java/ru/practicum/event/service/impl/PublicEventServiceImpl.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
-import ru.practicum.dto.event.EventFullDtoResponse;
-import ru.practicum.dto.event.EventShortDto;
-import ru.practicum.dto.event.EventState;
+import ru.practicum.event.dto.EventFullDtoResponse;
+import ru.practicum.event.dto.EventShortDto;
+import ru.practicum.event.dto.EventState;
 import ru.practicum.event.mapper.EventMapper;
 import ru.practicum.event.model.Event;
 import ru.practicum.event.model.QEvent;
