+++ ewm-main-service/src/main/java/ru/practicum/participation/dto/ParticipationRequestDto.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.dto.EventRequestStatus;
+import ru.practicum.event.dto.EventRequestStatus;
 
 import java.time.LocalDateTime;
