+++ ewm-main-service/src/main/java/ru/practicum/participation/mapper/ParticipationRequestMapper.java
@@ -1,7 +1,7 @@
 package ru.practicum.participation.mapper;
 
 import org.springframework.stereotype.Component;
-import ru.practicum.dto.EventRequestStatus;
+import ru.practicum.event.dto.EventRequestStatus;
 import ru.practicum.participation.dto.ParticipationRequestDto;
 import ru.practicum.event.model.Event;
 import ru.practicum.participation.model.ParticipationRequest;
