+++ ewm-main-service/src/main/java/ru/practicum/participation/service/impl/PrivateParticipationRequestServiceImpl.java
@@ -5,9 +5,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.dto.EventRequestStatus;
+import ru.practicum.event.dto.EventRequestStatus;
 import ru.practicum.participation.dto.ParticipationRequestDto;
-import ru.practicum.dto.event.EventState;
+import ru.practicum.event.dto.EventState;
 import ru.practicum.event.model.Event;
 import ru.practicum.event.service.EventService;
 import ru.practicum.exception.ConflictParticipationRequestException;
